Change `capacity` type from `Int` to `Double` across models and add ISO 8601 date formatting for `HistoryScreen`.

+++ composeApp/src/commonMain/kotlin/com/fruex/beerwall/ui/models/Models.kt
@@ -27,7 +27,7 @@ data class Transaction(
     val commodityName: String,
     val startDateTime: String,
     val grossPrice: Double,
-    val capacity: Int
+    val capacity: Double
 )
 
 @Immutable
